+++ pvi/Object.py
@@ -22,7 +22,7 @@
 
 from __future__ import annotations
 from typing import Dict, Any
-from ctypes import create_string_buffer, c_int32, byref, sizeof
+from ctypes import create_string_buffer, byref, sizeof
 from typing import Callable, Union
 import re
 import ast
@@ -196,64 +196,6 @@ class PviObject():
             self._linkDescriptor.update( {'EV': str(s)})
         else:
             raise PviError(self._result, self)      
-
-
-    @property
-    def refresh(self) -> int:
-        '''
-        refresh time [ms]
-
-        Typical usage example:
-        ```
-        temperature = Variable( task1, name='gHeating.status.actTemp', RF = 2000 )      
-        ```
-        '''
-        t = c_int32(0)
-        self._result = PviRead( self._linkID, POBJ_ACC_REFRESH , None, 0, byref(t), sizeof(t) )
-        if self._result == 0:
-            return t.value
-        else:
-            raise PviError(self._result, self)  
-
-    @refresh.setter
-    def refresh(self, time : int ):
-        '''
-        refresh time [ms]
-        '''
-        t = c_int32(time)
-        self._result = PviWrite( self._linkID, POBJ_ACC_REFRESH, byref(t), sizeof(t), None, 0 ) 
-        if self._result:
-            raise PviError(self._result, self)  
-
-
-
-    @property
-    def hysteresis(self) -> float:
-        '''
-        event hysteresis
-
-        Typical usage example:
-        ```
-        temperature = Variable( task1, name='gHeating.status.actTemp', HY = 5.0, LinkDescriptor="LT=prc" )        
-        ```
-        '''
-        s = create_string_buffer(b'\000' * 10)   
-        self._result = PviRead( self._linkID, POBJ_ACC_HYSTERESE , None, 0, byref(s), sizeof(s) )
-        if self._result == 0:
-            s = str(s, 'ascii').rstrip('\x00')
-            return float(s)
-        else:
-            raise PviError(self._result, self)  
-
-    @hysteresis.setter
-    def hysteresis(self, h : float ):
-        '''
-        event hysteresis
-        '''
-        s = create_string_buffer(str(h).encode('ascii'))
-        self._result = PviWrite( self._linkID, POBJ_ACC_HYSTERESE, byref(s), sizeof(s), None, 0 ) 
-        if self._result:
-            raise PviError(self._result, self)  
 
 
     @property